misc fixes and updates:
* lookupd: fix double write of response
* move GetTopicChannelArgs into its own util file and re-use in nsq_pubsub
* add Content-Length to ApiResponse(); use in a few more places
* cleanup TODO (added as issues)
* bump versions
