+++ TODO.md
@@ -1,7 +1,4 @@
 
 ### Roadmap
 
- * v0.3 nsqadmin - more complete administrative commands, UI for topology/stats
-    * upon topic creation, lookup channels against lookupd
-    * cleanup (expire) topics/channels
  * v0.4 working TLS for communication channel - usable between datacenters
