+++ nsqd/version.go
@@ -1,3 +1,3 @@
 package main
 
-const VERSION = "0.2.6"
+const VERSION = "0.2.7"
