+++ nsqlookupd/version.go
@@ -1,3 +1,3 @@
 package main
 
-const VERSION = "0.2.3"
+const VERSION = "0.2.4"
